+++ packages/auth/facebook/src/auth-facebook/auth-facebook.component.scss
