+++ packages/auth/facebook/src/public_api.ts
@@ -1,0 +1,2 @@
+export * from './auth-facebook/auth-facebook.component';
+export * from './shared';
