+++ packages/auth/facebook/src/shared/auth-facebook.interface.ts
@@ -1,0 +1,4 @@
+export interface AuthFacebookOptions {
+  enabled?: boolean;
+  appId: string;
+}
