+++ packages/auth/google/ng-package.json
@@ -1,0 +1,1 @@
+{}
