+++ packages/auth/google/src/auth-google/auth-google.component.scss
