+++ packages/auth/google/src/public_api.ts
@@ -1,0 +1,2 @@
+export * from './auth-google/auth-google.component';
+export * from './shared';
