+++ packages/auth/google/src/shared/auth-google.interface.ts
@@ -1,0 +1,5 @@
+export interface AuthGoogleOptions {
+  enabled?: boolean;
+  apiKey: string;
+  clientId: string;
+}
