+++ packages/auth/google/src/shared/index.ts
@@ -1,0 +1,1 @@
+export * from './auth-google.interface';
