+++ packages/auth/internal/src/auth-intern/auth-intern.component.ts
