+++ packages/auth/internal/src/public_api.ts
@@ -1,0 +1,1 @@
+export * from './auth-intern/auth-intern.component';
