+++ packages/auth/microsoft/ng-package.json
@@ -1,0 +1,1 @@
+{}
