+++ packages/auth/microsoft/src/auth-microsoft/auth-microsoft.component.html
