+++ packages/auth/microsoft/src/auth-microsoftb2c/auth-msalServiceb2c.service.ts
