+++ packages/auth/microsoft/src/public_api.ts
@@ -1,0 +1,4 @@
+export * from './auth-microsoft.provider';
+export * from './auth-microsoft/auth-microsoft.component';
+export * from './auth-microsoftb2c/auth-microsoftb2c.component';
+export * from './shared';
